+++ lib/change_health/response/claim/report/report_835_health_care_check_remark_code.rb
@@ -1,0 +1,11 @@
+module ChangeHealth
+  module Response
+    module Claim
+      class Report835HealthCareCheckRemarkCode < Hashie::Trash
+        property :code_list_qualifier_code, required: false
+        property :code_list_qualifier_code_value, required: false
+        property :remark_code, required: false
+      end
+    end
+  end
+end