+++ test/change_health/response/claim/report/report_claim_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class ReportClaimTest < Minitest::Test
+  describe 'report claim lists' do
+    let(:report_claim) { ChangeHealth::Response::Claim::ReportClaim.new }
+  end
+end